requirements.txt
absl-py==0.10.0
amqp==2.6.0
asgiref==3.2.10
astunparse==1.6.3
billiard==3.6.3.0
cachetools==4.1.1
celery==4.4.5
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
configparser==5.0.0
crayons==0.3.1
cryptography==2.9.2
dj-database-url==0.5.0
Django==3.0.7
django-appconf==1.0.4
django-cors-headers==3.4.0
django-cryptography==1.0
django-heroku==0.3.1
djangorestframework==3.11.0
future==0.18.2
gast==0.3.3
google-auth==1.21.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.32.0
gunicorn==20.0.4
h5py==2.10.0
idna==2.10
importlib-metadata==1.6.1
Keras-Preprocessing==1.1.2
kombu==4.6.10
lxml==4.5.1
Markdown==3.2.2
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==1.0.5
pandas-datareader==0.8.1
pandas-market-calendars==1.5.0
protobuf==3.13.0
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
selenium==3.141.0
six==1.15.0
SQLAlchemy==1.3.17
sqlparse==0.3.1
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
urllib3==1.25.9
vine==1.3.0
webdriver-manager==3.2.1
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
zipp==3.1.0
